addressed issues
Updated descriptions and renamed stuff based on issues pointed out by Jenny and Vinny.

This big (annoying) one is going to be naming all the variable with lowercase letters. I had used LJ for a few things, which got changed to lj. (sLJ -> s_lj is the only exception to this)