Added LJ example
Added the basics steps on how to use the package.

Feel free to play around with starting points ([variable]_init) in the model definition or ranges ([variable]_lower and [variable]_upper). Notice for LJ model if you start sigma too far from the true value the MCMC can't find the true values. (this is because sigma is ^12